Fix: 로그 데이터 Integer로 수정, 데이터 없어도 0으로 나오게 수정

+++ backend/log-service/src/main/java/com/helloword/logservice/domain/log/repository/LogRepositoryCustom.java
@@ -8,9 +8,9 @@ public interface LogRepositoryCustom {
 	Integer findTodayTotalPlayTime(Long kidId);
 	Integer findTodayCompletedGameCount(Long kidId);
 	Map<String, Integer> findDailyCorrectWordCount(Long kidId);
-	Map<String, Double> findGlobalDailyAverageCorrectWordCount();
-	Double findKidAverageCorrectRateByGameType(Long kidId, GameType gameType);
-	Double findGlobalAverageCorrectRateByGameType(GameType gameType);
-	Double findKidAveragePlayTimeByGameType(Long kidId, GameType gameType);
-	Double findGlobalAveragePlayTimeByGameType(GameType gameType);
+	Map<String, Integer> findGlobalDailyAverageCorrectWordCount();
+	Integer findKidAverageCorrectRateByGameType(Long kidId, GameType gameType);
+	Integer findGlobalAverageCorrectRateByGameType(GameType gameType);
+	Integer findKidAveragePlayTimeByGameType(Long kidId, GameType gameType);
+	Integer findGlobalAveragePlayTimeByGameType(GameType gameType);
 }
